internal/testing/integration: add  internal.ExperimentUnitPage to tests
internal.ExperimentUnitPage is added to tests, so that the integration
tests are testing the unit page code path.

This is necessary to remove ExperimentFrontendDocRender in the next CL.

Additionally, tests for autocomplete in TestEndToEndProcessing are
deleted to simply the code and because autocomplete isn't being used.

Change-Id: Iffa0205b1b4a77a048e71c0a93c8f08f60af9365
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/272606
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>