internal/frontend: fix pageTitle for v2+ module versions
At the moment, the page title for a module version with a major version
of v2 of higher is the major version. It is now changed to the last
element of the module path.

For example, the title for mvdan.cc/sh/v3 is now sh instead of v3.

Change-Id: Idaf26bb87d7abb8ab2f8a7a15c8641d089ccdb25
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/275298
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>