all.bash: update prettier script to use dockerized prettier install
Updates all.bash to use a dockerized prettier install as a fallback
when prettier is not installed globally. For now the script will
only use the container if it has already been built. We could change
this in the future to automatically build the image but this will give
everyone a chance to test it out first. To build the image run
'docker-compose -f devtools/config/docker-compose.yaml build' from the
repo root.

Change-Id: I56ad2cd2da2fe0dbbf46636024f44cc614adb168
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/275532
Trust: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>