internal/worker: record processing lag metric
Create a metric for the age of the oldest unprocessed module, and record it periodically.

We choose to record it whenever the /poll endpoint is hit. Since that
endpoint is hit periodically (once a minute) by Cloud Scheduler, we
can be sure that we will get exactly one data point per minute. The
downside is that we won't have finer granularity than that, but since
the main purpose of this metric is to detect large lags, that doesn't
matter.

Change-Id: I7ce3f043cceaf567fb0c871676063eda91bb2420
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/278614
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>