internal/licenses: upgrade to latest licensecheck
The new licensecheck is a rewrite of the old with a different API
and implementation. We will eventually switch to it, but first we
upgrade to a version with both the old and new APIs, and continue
using the old one.

I confirmed (by running an internal tool) that the change to
github.com/google/licensecheck/old@v0.3.1 doesn't affect any licenses
in our database.

Change-Id: I825b139b45c05b711c39b6018d7ba5f89a1886ad
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/281952
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>