internal/licenses: use new licensecheck
Use the new version of licensecheck, which is more precise and has
other improvements.

The new version uses regexps instead of fuzzy matching to categorize
licenses. That makes it much stricter than the old version,
so we need many exceptions in order to match the same licenses
we did before.

It is also intolerant of various corruptions like truncation
and search-and-replace errors. We observed many of these,
and filed bugs against the offending repos where possible.

We validated that except for corrupted licenses, this package accepts
the same licensesn as it did before for all latest-version modules.

The new version uses SPDX identifiers where they exist, so we no
longer need to map licensecheck names to SPDX IDs. We also need
include many variant SPDX identifiers in our list of redistributable
licenses (e.g. "GPL2.0-or-later").

The new version dispenses with "-Short" and "-Header" variants,
so we don't need to canonicalize the names.

Change-Id: I5c1aaa2bdf745aff6ebb8c121fa63d6550931069
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/283654
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>