internal/frontend: refactor resultFromFetchRequest
fetchRequestStatusAndResponseText is renamed to resultFromFetchRequest,
and now returns a fetchResult and error.

This refactor will allow us to use this function on a 404 in the next CL
in order to determine if the result of a previous fetch can be shown.

When a previous fetch request results in a 500, allow a refetch of that
path when the taskIDChangeInterval has exceeded.

Change-Id: I978eeccaa2f58c9e640f25b8773b4ecfab7e9717
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/281676
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>