internal/frontend: return previous fetch result on 404
When a request results in a 404, show the user the previous fetch
result, instead of giving the option to refetch the path. The majority
of the time, clicking "Request <path>" will just show the user when is
in the database.

Change-Id: I652edd86e8ad236498eb78ffc205171309cbbee1
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/281675
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>