internal/postgres: save and restore new licensecheck Coverage
Add a NewCoverage field internal/licenses.Metadata to support
the upcoming change to the new licensecheck.

When saving to the DB, save the NewCoverage field if it is populated;
otherwise save the original Coverage field.

When reading the JSON for the coverage field from the DB, distinguish
old and new Coverage structs and populate one of the two fields of the
licenses.Metadata value as appropriate.

Change-Id: I5be5623fbaec668265e1e30a878864cf4aef2ae6
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/283652
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>