internal: resolve default branch versions for each fetch
The target for @main and @master is maintained in the version_map table,
while @latest and a specific semantic version is determined using the
modules and units tables. As a result, these versions can be out of
sync.

Resolve the master/main version when a module is fetched and
disableProxyFetch is off, so that the targets are up to date with the
new version.

Fixes golang/go#43096

Change-Id: I86dde167639c0f180ab67b0d3243316240956fc4
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/286212
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>