internal/stdlib: fix getGoRepo for requests to master
getGoRepo currently looks for a "master" tag when fetching the master
repo, instead of at the master branch. This is now fixed.

However, stdlib.semanticVersion still returns "master" when the
requestedVersion is "master", instead of returning a resolved semantic
version, so a fetch request for std@master will still result in a 500. A
comment is added to fix this in the future.

In a future CL, a fake pseudoversion will need to be constructed in
order to resolve std@master to a semantic version.

For golang/go#43890

Change-Id: I9e68b91d8c9e0bea6f6751d834853b852f0d532b
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/286153
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Trust: Julie Qiu <julie@golang.org>