internal/godoc: fix some bugs in example rendering
First, don't try to render an example if we can't convert the AST to a
string. (This was causing a panic.) Instead, render a message saying
that we failed to produce an example. In practice, this is only going
to happen for playable exammples, because we synthesize the AST for
those and that is a tricky proposition. We will log an error if that
happens, so occasional error sweeps of the logs will find these, but
absent those we will have to rely on user reports.

Second, fix the bug that caused the panic in the first place, which
was that the code for trimming unused global declarations could
convert a valid const group like
```
const (
	a = iota
	b
)
```
to merely
```
const b
```
which isn't valid Go. The fix is to keep an entire iota const group if
any of it is used.

Change-Id: I33637aa1e88d9c8731bb2ccc58e6991b84eba6b6
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/286914
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>