internal/frontend: fix resultFromFetchRequest on 200
When a fetch request returns in a 200, it is possible that the fullPath
that is requested is not in that module. In that case, return the error
message for when we find a module path matching the full path prefix,
instead of a status 200.

Change-Id: I336ca3b260874caf7455eb620a20d9ef69c659e8
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/287292
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>