internal/frontend: show a page for an empty directory
Currently, it returns 404 for an empty directory.
However if the directory has nested modules, it would be useful
to show a page containing them.
So this change is to redirect requests for an empty directory
that is above nested modules to the search result page.

Fixes golang/go#43725

Change-Id: Icb0aae8c857f44931a999a74d45b29aa12a0c56c
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/290049
Reviewed-by: Julie Qiu <julie@golang.org>
Trust: Jonathan Amsterdam <jba@google.com>
Trust: Julie Qiu <julie@golang.org>