internal/database: catch missing retriable errors
We've noticed that we aren't retrying errors that have the right
serialization code. The likely reason is that the error type is not
what we expect.  Get information about the type so we can recognize it
in the code.

For golang/go#43899

Change-Id: I72aceb12002eecbf580654b77c68d845759e8182
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/291609
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>