devtools/config: create docker-compose.next config and run script
The config creates the following services
- nodejs: used to run npm commands, a slight refactor of the
existing config for npm in docker-compose.yaml
- ci: extends nodejs, used to run the e2e tests
- frontend: runs the frontend server
- migrate: runs database migrations
- wait_for_db: a utility service, used to delay the start of the
migrations until the db is ready for connections
- db: exposes a postgres database

Using the devtools/docker_compose.sh script we can run these
services and perform the relevant cleanup. The script exposes
the nodejs and ci services. We'll use it in all.bash and CI
to run the linters and tests for frontend assets.

For golang/go#43361

Change-Id: Ie9c91058f9c6796b737e94c237b2d99f612f45a7
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/292695
Trust: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>