internal/queue: remove CreateTask timeout
Remove the a timeout on the CreateTask API call. It was causing
failures on the production worker.

A comment said this was a requirement of the Tasks API, but I couldn't
find that in the docs. In any case, if it is, then presumably
CreateTask will return an error.

Also, add a stack to the returned error, for better error reporting.

Change-Id: Id40e596cafade6d673c1787fe446bfdff1a8a62a
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/293609
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>