internal/fetch: functions for fetching raw latest version
Add RawLatestInfo, which uses the proxy to find the raw latest version
of a module and its associated go.mod file.

The fetchRawLatestVersion function implements the full definition of
"raw latest" as embodied in the current (go 1.16) go command.  In some
cases this requires looking for the presence of a go.mod file.
Although we could always download the zip from the proxy, the function
accepts a function argument that can avoid that. The worker will pass
a function that looks in the database.

For golang/go#44437

Change-Id: I84e19a4a7a10a87d5664217e0404bc795b3f9470
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/295429
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>