internal/worker: update raw latest info on each fetch
Every time a module is fetched, update its raw latest information in
the database to the most recent value.

This is a simple way of making sure the DB has the most recent raw
latest info.  In the steady state it's unlikely to be wasteful, since
most of the time the new module version will be the latest, so the
update is needed.

For golang/go#44437

Change-Id: If64bca41b51fe63fb2aeb836d6ed98543090cd18
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/295890
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>