internal/postgres: insert paths as a separate transaction
When inserting a module, move inserting paths to a separate
transaction that is done first.

We can then lower the isolation level of the main transaction, because
we only increased it to RepeatableRead to deal with insertPaths. See
the commit message for aada664a381f020c365ff8a12c6b0e0f9f7c1cb0.

The hope is that we'll avoid deadlocks and other problems we've seen
that all seem to be issues with the paths table.

Change-Id: I7840b87ace9237cb0f8ea6ff888f2a509034ba0b
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/296929
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>