content,internal: add build context dropdown to unit page
Adds build context of the documentation to the unit page.
When there is a single buildcontext and it is linux/amd64
nothing will be displayed. A single non-standard build
context will show "Rendered for <build context>"
and multiple build contexts will show "Rendered for"
with a select element to switch the current build context.

Fixes golang/go#37232

Change-Id: I06ddbced9a444f3d5af19b189c509abc3241e4dc
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/297552
Trust: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>