internal/version: implement full latest logic
Implement the full latest-version logic in internal/versions, because
we'll need it in a couple of places. Also generalize it to a single
list of versions that may include pseudo-versions.  That way it's not
tied to the proxy's two endpoints (list and @latest) and can be used
on any list of versions, like a list of all module versions.

For golang/go#44710

Change-Id: I82ad9ae1e585762d13fecb1997c89312b88ad54d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/297889
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>