internal/database: don't report errors in logQuery
Remove the call to the error reporting service in logQuery.  It is at
too low level: in transactions that are retried, it can report an
error even though the error is a serialization failure that is normal
and will cause a retry.

Change-Id: Ia4d4ab354f68bff237adc93a5cf61ff7448a05af
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/298509
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>