migrations: use expression index on package_symbols
The UNIQUE INDEX on
package_symbols_package_path_id_module_path_id_section_synopsis_key is
changed such that the uuid of synopsis is used instead of the synopsis
string.

This is because the synopsis can be very long, whereas
uuid_generate_v5(uuid_nil(), synopsis) is fixed at 128 bits.

In CL 298769, this was attempted using a sha256. However, this fails on
strings when handling escape strings. CL 299329 was the revert.

Change-Id: I3102c32bcb5e54a64a44ef88c111a36300da83ea
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/299351
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>