internal/postgres: improve GetPackageSymbols query
The GetPackageSymbols query is changed such that the package_path
and module_path is determined by comparing against modules and units,
instead of package_symbols. This eliminates a second join on the paths
table, and is faster than when filtering on the join with
package_symbols.package_path_id and package_symbols.module_path_id.

For golang/go#37102

Change-Id: If75c81a399f6273ef814c8ca7b1495b20ea47d91
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/299889
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>