migrations: change package_symbols unique index
The UNIQUE index for package_symbols is changed to
(package_path_id, module_path_id, symbol_name_id, section, synopsis).

Within a given package, a symbol's parent or type should not be able to
change without the symbol's name, section, and/or synopsis changing.

However, the section can change with the name and synopsis staying the
same. For example, if `const Foo` is in the Constants Section, but is
later modified such that it is of type T, it will now have parent T and
and section Types, instead of parent Foo and section Constants, but the
name will be Foo and synopsis will be `const Foo` in both cases.

For golang/go#37102

Change-Id: I3f8bb84d7376c1f9d0ebbd85c82329b4f27fa7f3
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/301050
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>