internal/postgres: upsert search with module path and version parameters
When we upsert a row in search_documents, we know the package, module
and version.  Use that information to select the information to
upsert, instead of finding the latest version with an ORDER BY clause.

Besides being unnecessarily inefficient, the ORDER BY is not accurate
in the presence of retractions and other adjustments.

Change-Id: I49a96add12ede3e909d70b5e55094c1b704a1822
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/301510
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>