devtools,internal: use dynamic import for dialog polyfill
Loads dialog polyfill with a dynamic import. This ensures
that the polyfill is explicity loaded where required and
that it does not load when there are no dialog elements
on the page. Additionally this will prevent a race condition
in some browsers where a dialog might be registered before
the polyfill JS has been parsed.

Change-Id: I7980ca62efd3bca9ad78422973d92706bc8ff4a5
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/302010
Trust: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>