migrations: add documentation bigint
The rate of growth for the documentation table suggests that the id
column might hit the INTEGER limit as the size of our database grows. We
change this to a BIGINT column now, since its easier to make this change
before this table is used in prod.

The update_documentation_id is changed to update both documentation.id
and documentation.id_bigint with the same value on INSERT or UPDATE.

Change-Id: Id310ed548c2fa3769fbd2707cea4f69295434f2f
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/302671
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>