content,internal: readme outline fixes
Heading nodes that contained only non-text content
would show as whitespace in the outline. This change
updates the outline generation logic to use text
extracted from non-text nodes when a heading contains
no plaintext.

A first pass through nodes in a heading looks for text
and writes found text to a buffer. If the buffer is
empty after the second pass, other nodes are used to
generate the heading label text.

Text breaking needed for documentation causes issues with
text in tables in the readme. Moved the word break styles
to the UnitDoc class.

For golang/go#45164

Change-Id: I0fa005d43c605f972d669966c80369c4fdaea85d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/303749
Trust: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>