internal/frontend: extract nested table of contents from markdown
The processReadme function now returns a nested table
of contents extracted from the readme file based on
the hierarchy of headings in the document.

If the document has only a single top level heading
and that heading has children it is assumed to be a
title and removed from the table of contents.

For golang/go#43325

Change-Id: I5b67a7e356820477851e6c34493e2517f889b02f
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/309393
Trust: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>