content/static: fix uncaught error in dialog-polyfill
Pressing the enter key to select an item from
the jump modal causes an error because the default
behavior of the modal is to close on submit. We already
close the modal when the item is clicked, so we can
prevent the default behavior in this instance.

Change-Id: I91ab270aafdb77b3680bf6bbac23144fe3e263e3
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/313511
Trust: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>